adc_python_client/requirements.txt
cx-Freeze==6.4.2
importlib-metadata==3.1.1
numpy==1.19.2
Pillow==8.0.1
pypiwin32==223
pyserial==3.4
pywin32==228
six==1.15.0
win10toast==0.9
wxPython==4.1.0
zipp==3.4.0
